requirements.txt
python-telegram-bot[job-queue]
SQLAlchemy>=2.0
asyncmy
alembic
zai-sdk
openai
google-genai
binance-futures-connector
aiohttp
telegramify-markdown
python-dotenv
